+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Unreleased
 ----------
 
 -  feat: :doc:`/scripts/csvcut` adds an :code:`--ignore-unknown-columns` option to skip identifiers in :code:`-c/--columns` that do not match a column in the input.
+-  feat: :doc:`/scripts/csvclean` adds a :code:`--remove-empty-columns` option to remove empty columns from standard output.
 
 2.2.0 - December 15, 2025
 -------------------------